rwdt: protect RWTCNT write sequence with spinlock
It is necessary to ensure we don't write RWTCNT while the WRFLG is 1
in RWTCSRA, or the device can misbehave.

Protect the "check RWTCSRA then write RWTCNT" block with a spinlock
to avoid SMP/multithreading issues with another core writing RWTCNT
after our RWTCSRA read.

(This assumes that WRFLG only transitions from 0 to 1 as a result of a
write to RWTCNT - otherwise we could never guarantee safety.)

SSGLOGEO2-1422 / PCP KB13070835196

Signed-off-by: Kevin Bracey <kevin.bracey@renesasmobile.com>