rwdt/cmt15: use jiffies to rate-limit touches
Rather than read the hardware counters (slow), use jiffies to rate-limit
touches.

This version also moves the time check from the core RMU kick routines
to the touch_softlockup_watchdog API, making the core kick routines
unconditional again. This is probably safer, as it will avoid problems
with important kicks not happening, as we've seen happen.

The previous version didn't write to the hardware if the counter was
less than 1/8 through the expiry time. Now, the touch routine does not
call the core kick routines if (according to jiffies) a touch or
periodic kick occurred within the last 1/4 of a periodic kick time.
(Other direct kicks using rmu2_rwdt_cntclear or rmu2_cmt_clear are not
considered).

Note that if IRQs are not being handled, jiffies won't be incrementing,
and as a result, we may end up not touching the watchdog. This partially
reduces the efficacy of the touches - a long run of serial console
output printks calling touch with IRQs enabled will keep us alive, but
the same with IRQs disabled will allow the watchdog to trigger. But this
doesn't seem a totally unreasonable result.

PCP KB13070835196

Signed-off-by: Kevin Bracey <kevin.bracey@renesasmobile.com>