rwdt/cmt15: add touch loop tests
Add test modes 12 and 13 to test infinite loops with calls to the
watchdog touch routines.

Also note that TEST_LOOP_ALL (all CPUs IRQ-enabled loop) is
unimplemented - it's not straightforward to implement in the current
framework - on_each_cpu calls functions with IRQs disabled.

Tidy up the formatting of the test mode enum - it was mangled in an
attempt to make it less than 80-characters wide.

Signed-off-by: Kevin Bracey <kevin.bracey@renesasmobile.com>