memmap change: make PM code cope (TZ only)
Modify the PM code to handle the memory map change. The PM code has
sections that run with the MMU on, and sections with the MMU off;
fortunately, there was infrastructure in place to support the
potentially different mappings, but this hadn't been exercised, and
needed some work.

To make the MMU enabling code safe, we need an identity mapping for the
area its running from - if we were running in normal SDRAM, we could
just put that code in the idmap area. As we're not (why?), I've hacked
the core idmap code to include the RAM0 area, as a simple solution for
now.

I've only tackled the TrustZone version - the equivalent set of changes
will be required for the non-TrustZone version.

I also note that the disabled PASR_SUPPORT code lacks any proper
handling of mapping for its hardware accesses - I haven't fixed this up,
and the code won't compile at present.