memmap change: mark PM functions in ICRAM executable
Since the memmap change, aborts have been seen on the transition from
SDRAM to ICRAM in the PM code. The aborts appear to be permission
faults, probably related to some sort of synchronisation failure on the
domain switch. The ICRAM is marked XN (eXecute Never), and the domain
switch should override that.

Not clear what the synchronisation problem is, or why the memmap change
has caused it - documentation and Linux code suggest a simple ISB after
the domain control register write change should suffice.

But as it's an ugly hack anyway, remove the need for it by arranging for
the relevant section of ICRAM to be marked as MT_MEMORY_NONCACHED rather
than MT_DEVICE, and hence be executable.

Note that this changes the ARM mapping type from Device to Normal. (The
Linux code doesn't have a MT_xxx entry for executable Device memory, and
anyway fetching instructions from Device memory is UNPREDICTABLE
according to the ARM Architecture Reference Manual).

The ICRAM has restrictions on access size which means we should probably
avoid the use of a Normal memory mapping rather than Device for
read/write data areas, but Normal appears to be work for the code area -
we only write in one place (copy_functions), and cache-line-fill reads
must already be supported, as MMU-off instruction fetches act as Normal
memory.

One minor impact is that the HPB IO space is no longer covered entirely
by a section mapping - the range E6200000-E63FFFFF is now page-mapped.
This will impose a slight amount of extra TLB load.