memmap change: Remove most _PHYS defines; add IO_TO_PHYS()
There is no need to have virtual and physical defines for every single
register. Rationalise to only have virtual defines for the vast majority
of registers in the E6xxxxxxx space.

In the process, restore <mach/r8a7373.h> closer to the
pre-memmap-change code, to try to reduce conflict/porting issues.

Add an IO_TO_PHYS macro as the opposite of IO_ADDRESS, to get the
physical address of a virtually-defined register in the E6xxxxxx and
F0xxxxxx spaces.