r8a7373: remove unused clk pointer cmt10_clk
Disabling and enabling cmt10_clk do not do anything because cmt10_clk
pointer is not set at all. Setting the cmt10_clk was removed at the same time
when cmt timer driver was changed but git commit message does not say why.

I tried setting the pointer but after that the device never got in to suspend
but freezed before that so I cleaned out disable and enable that clock.