CONFIG_GIC_NS_CMT - remove it completely - JK13090559973
We have a strange configuration called CONFIG_GIC_NS_CMT in
the watchdog related files. This is actually harmful, since it was
not on for example in KOTA and we had to find out the hard way why
we were getting lots of SW Watchdog resets when removing the printk()
from watchdog kicking.

Turns out all the CMT watchdog kicks are actually behind this
flag (it's not even a real config, it's just a #DEFINE in the files).

PCP reference: JK13090559973

(cherry picked from commit 894303cb63b6539308ceee9476ac8a483233eed4)

Conflicts:
	drivers/misc/rmu2_cmt15.c