Support for i2c over device tree
Added i2c bus nodes in the DT. No major i2c bus driver modifications needed, but
the clock had to duplicated for DT. In a non-DT setup, the i2c clock name was
taken as i2c-shmobile.X, but in DT the clock names is taken as address.i2cX,
where address is given in the i2c node. The driver was modified to accept the
'clock-frequency' attribute.

i2c buses where tested by probing using the i2cdetect tool. Currently i2c-0,
i2c-3 and i2c-4 are working fine. Other buses timeout. DT for i2c-8 is not
supported by the driver (its a different driver, sh7730).

TODO: fix time-outs in 1, 2, and 7 buses (5 and 6 are not used, but is worth
checking it too.)

Signed-off-by: Asha MVM <ext-asha.magadi-venkatesha-murthy@renesasmobile.com>
Signed-off-by: Santosh Sivaraj <ext-santosh.sivaraj@renesasmobile.com>