Fix 3.0->3.4 merge errors
This commit fixes up a number of merge errors that occured in the
previous Linux 3.0->3.4 merge. These were discovered while auditing the
code during the 3.4->3.10 merge.

Most changed are in drivers we aren't using, and others are things like
removing little fragments of duplicate code - the same error check being
done twice, for example. Most files affected now precisely match the
upstream code in commit 54ae97fb on the android-3.4 branch; others just
get closer to it.

The only change that looks like it could actually have a potential
impact in code that we're using is the removal of a sync call in
mm/vmalloc.c; this comes from commit cd12909cb, in Linux 3.2.