Revert "ARM: Make low-level printk work"
This reverts commit 4f97a420c5e1c7beff2ebc8d69fa439a448d1eca from the
Android branch.

It's not clear what purpose this commit serves - it means that all
printk output is sent to both the DEBUG_LL channel and the normal
console, so everything is output twice if DEBUG_LL is enabled and your
DEBUG_LL channel is the same as your normal console (as it is for us).
The only benefit I can see is that you get early printk without changing
your command line.

Without this commit, earlyprintk functionality works fine if you add it
to your kernel command line, and the system performs the expected
early->normal printk handover. Nothing is printed twice, apart from a
line or two at the instant of handover.

Signed-off-by: Kevin Bracey <kevin.bracey@renesasmobile.com>