r8a7373-loganlte.dts: correct DT memory size to 880M.
Memory size was set to 512M, copied from the unused CONFIG_MEMORY_SIZE.
Increase to be the full 1G minus the start offset = 880M.

Doesn't have any real effect in the current build, which has
CONFIG_ARM_ATAG_DTB_COMPAT on, which means we get the full memory from
the boot loader ATAGs, but gets the memory right with that off.

But I note that there is (and already was) a double counting in the boot
logs - it says "Memory: 880196k/1760392k available", with the second
number being the memory declared in the DT plus the memory declared in
the ATAGs, not accounting for the overlap.