loganlte: set more D2153 LDOs to "always_on"
When using device tree, "regulator_has_full_constraints" is assumed to
be true. This means that the regulator core will do a post-init
clean-up to disable all zero-reference-count regulators.

To survive this in the current setup, we apparently need to set 3 more
LDOs to always_on - 1 (VDIG_RF), 4 (VVCTCXO) and 5 (VMIPI).