ARM: shmobile: r8a7373: reactivate cascaded IRQs for DT
Reactivate the code that provides the external interrupt sources for
device-tree builds.

The extra IRQ controller implementation is non-DT, and assumes that we
the DT GIC node is present and correct.

There is a new upstream irq-renesas-intc-irqpin driver, which we could
look at using instead, but we may need to add some functionality like
debounce control.

With this patch and the preceding ones for I2C bus numbering and
always_on modifications, the D2153 regulator driver now works in
device-tree builds.