sh: reinstate pm_runtime.c
On our 3.4 branch, pm_runtime was removed from drivers/sh/Makefile,
and this removal was taken in the new 3.4->3.10 merge.

The reason for the change is not fully understood; we believe it's
an interaction with CONFIG_PDC.

On the current build, with pm_runtime absent, pm_clk_add_notifier is not
called, so drivers/base/power/clock_ops.c does not automatically turn on
clocks for drivers based on the CLKDEV_CON_ID tables. And unused clocks
are turned off after boot by clk_late_init (something which was not
present in our 3.4 branch), all of which means the SYS-DMAC clock is
gated off, and DMA doesn't work.

Reinstate pm_runtime.c if !CONFIG_PDC, to get standard Linux clk
enabling behaviour, so the shdma driver gets its clock. We will need to
revisit this if/when CONFIG_PDC is reenabled.

As of this commit, eMMC is now working with DMA.