crashlog: adapt crashlog to 3.10 logcat logger
in 3.10 the logcat log buffer are created dynamically so old way of getting log buffer
info did not work. Also 3.10 uses vmalloc to allocated buffer and that had to be changed
to kmalloc to get physically contiguous buffer for logcat. it is needed to be able to split
logcat logs from the ramdump binary without trace32.

This commit also removes old IRQ trace log which is not used anymore to be able to
compile the file. That is already cleaned in k3.4.5 branch so it is safe to drop it also here.