sh-pfc: fix up hwspinlock support
Reimplement hardware spinlock support, which was originally implemented
in commits 9426e6da1 and 60667d704, and not fully merged across to 3.10.

The hardware spinlock to use is now obtained via a sh_pfc_soc_operations
call to the pfc-xxxxx.c file.

The hwspin_lock call is moved down to the appropriate core place, and is
thus always claimed inside the original pfc->lock spinlock, avoiding the
problems arising from the "if/else" claiming of the software or hardware
spinlock that 60667d704 had to fix.

No implementation in this commit - this is just the core framework.