pfc-r8a7373: use gpio data set/clear registers
For atomic manipulation of GPIO outputs, write to the the GPIO data set
and clear registers, rather than performing read/modify/write on the
data registers.

This commit simply adds the register definitions to the pfc-r8a7373.c
tables, thus activating the existing support in the core code.

This resurrects functionality lost in the 3.4->3.10 merge and initial
platform bringup.