pfc-r8a7373: Various table corrections
In pinmux_pins[], pins 133-136 and 139 did not have the correct IO/PU/PD
flags. Minor impact - 139 would not have been useable as an input, and
133-136 could have been erroneously set to output.

In pinmux_data_regs[], correct the names of the data registers. Not sure
if there's any visible effect.

In pinmux_irqs[], change from enum_ids to GPIO numbers, which makes
GPIO->IRQ mapping work.

In r8a7373_portcr_offsets[], add the missing final entry, which avoids a
crash when attempting to modify the bias on pins 320-327.