pfc-r8a7373: add hwspin_lock protection to set_bias
GPIO hardware semaphore protection is required when changing bias, so
add this protection inside r8a7373_set_bias.

Arguably, the hwspin_lock calls could be put in the core code, but as
the core code is calling down to the driver anyway, this leaves it up to
the individual driver whether a hwspin_lock is required for each op.
(Compare the set_debounce op, where it isn't).