r8a7373: ensure IRQC clocks are enabled
MSTP507, which gates the IRQC clock, was declared but unreferenced; this
meant that it got shut down after boot.

Add clock mappings for the renesas_irqc driver, although we don't use it
(yet), and make our local intc-r8a7373 irqchip code claim the
renesas_irqc.0 clock.

To permit access to the clock framework, the IRQC initialisation needs
to be moved later to init_machine, rather than init_irq. This is
consistent with where the IRQC init happens on other shmobile platforms,
and is where it would be if we do move to the renesas_irqc driver.