r-mobile_ion.c: restore prototype of r_mobile_ion_client_create
For backwards compatibility with out-of-tree modules, put back the
heap_type_mask parameter in the prototype of r_mobile_ion_client_create.
This parameter is ignored, as it is no longer provided by
ion_client_create (a change present in the android-3.10 branch we're
based on, and in later versions of android-3.4).

Signed-off-by: Kevin Bracey <kevin.bracey@renesasmobile.com>