checkpatch errors and warning removal for camera change.
removed errors and warnings in the boardfiles.
warning for extern inclusion in the setup-u2camera.c is not
removed as other alternative needs to be found. As of now without
this change the build breaks as boardfile compiles later than
setup-u2camera.c file. now we have data to be accessed from
boardfile in setup-u2camera.c before it is compiled.placing of this
extern in setup-u2camera.h is breaking other file comliation.

Signed-off-by: Preetam M N <Preetam.MN@renesasmobile.com>