proc entry fixes for fm32_we295 driver
Another proc entry fix. This module is required for android
compilation. Functionality is not tested yet. Removed debug logs that just
prints clock registers, which shouldn't be done from here. Protected the
registration with soundpath logical.

If android 4.4 doesn't support modules (??), then this needs to be changed to
obj-y and made corresponding changes to makefiles in vendor directory.

Signed-off-by: Santosh Sivaraj <ext-santosh.sivaraj@renesasmobile.com>