shmobile: logan: Fix secure side hang due to change in spinlock implementation
A spinlock is shared between secure side and public side for l2 cache access.
In the older spin lock implementation, A zero value for the lock always meant
it is avaliable/free for locking. On 3.10 - the implementation has changed.
A zero value values does not neccessarily mean it is available/free.

Hence on 3.10 calling 'l2x0_init_later' makes the secure side to hang
since it keeps spinning on the lock waiting for it to be unlocked.
This results in a secure side abort/reset.

This is a temporary workaround until the secure code is changed or the
spinlock implementation is changed.

We tested idle for over 4 hours and we did not observe any crash.

Signed-off-by: Sriram Venkatakuppaiah <ext-sriram.venkatakuppainah@renesasmobile.com>