Socket numbering changes to be compatible with the user space!!
The current build system does not use the current kernel headers, which should
be fixed. Live kernel headers should be used, rather than keeping an old copy of
the kernel header and fixed that in the new kernel.

TODO: Revert once when the build scripts are modified to use live headers

Signed-off-by: Santosh Sivaraj <ext-santosh.sivaraj@renesasmobile.com>