RTDS: Flush L1 cache during mmu flush
Lot of RT-Domain crashes have been seen in the system. On debugging we
see that there is a translation fault on RT-Domain despite having
correct mapping on ARM. Translation fault is seen 100% on playing video.

On flushing full L1 cache on during MMU flush (where L2 is flushed) is
avoiding this crash. This would be done currently as a workaround. The
full performance impact of this patch is not yet measured. This is put
to ensure we unblock all other modules to continue testing.

Video playback works with patch!!!

Signed-off-by: Naga Chandra Prahalada Rao <nagachandra@broadcom.com>