ARM: hw_breakpoint: revert to unmodified 3.10 version
Unwind the local modifications that were originally performed against
Linux 3.0 in commit 1fbda092, and have since been merged through 3.4 to
3.10.

The check of DBGREG1 before accessing debug registers is now
unnecessary, as since 3.2 the upstream code now copes with them being
unavailable; (except for the OS save/restore check added in 3.9,
which will be fixed in a following patch).

The separate arch_hw_breakpoint_init_late implementation for pmdbg was
inelegant, and is by now out of sync with arch_hw_breakpoint_init. This
will be reimplemented in a following patch as a repeat call of
arch_hw_breakpoint_init.