ARM: hw_breakpoint: export arch_hw_breakpoint_init
Make arch_hw_breakpoint non-static, so that it can be called to kick
hw_breakpoint functionality into life if debug hardware is unlocked
after boot.

This is unsafe if a previous init has succeeded, and may not be totally
safe even then, but it's only intended as a debugging tool.