ARM: hw_breakpoint: catch exceptions in OSSR check
The read from DBGOSLSR in core_has_os_save_restore() should always be
permitted according to the ARM ARM.

However, as described in erratum 764319, the Cortex A9 generates an
undefined instruction exception when DBGSWENABLE=0.

Cope with this by moving the call to core_has_os_save_restore() into the
block with trap handling.

Signed-off-by: Kevin Bracey <kbracey@broadcom.com>