shmobile: Fix random secure side watchdog resets
The commit 'bf53acdf23995b25c7583db2e4f773655efeb868' introduces a
a random secure side watchdog resets. The resets happen when
core0 is in corestandby2 while hotplug is in progress. This
condition is invalid. core0 should not enter corestandby2 when
hotplug is in progress.  This patch fixes this bug.

Signed-off-by: Sriram Venkatakuppaiah <ext-sriram.venkatakuppainah@renesasmobile.com>