bcm-bt-rfkill: convert boards to use DT
As an initial simplest-possible example of using the pfc gpio-controller
node, convert the boards to use DT setup for the bcm-bt-rfkill device.

Note that the setup for this device is identical for all 3 boards, so we
continue to get away with having only 1 shared DTS file - this will need
to be sorted out in the build before going much further.

Conditionality in the board files is a pain - I had to add ifndef
CONFIG_OF in about 16 places. That really just reflects the sorry state
of the board files, plus the need to support non-DT builds :( - it will
be better after Kari's clean-ups, at least.

Compensate for this by removing the conditionality on
CONFIG_BCM_BT_RFKILL in the board files - it's not necessary to have
such conditions cluttering them. Those configs indicate whether or not
the driver is built-in, but the hardware always exists or not on a given
board, regardless. (Consider the possibility of a module driver, built
later).

The driver itself needs no modification, as it already has DT support,
coming from Broadcom's other projects.

Signed-off-by: Kevin Bracey <kbracey@broadcom.com>