multiproduct cleanup: remove board-<name>.h
These files were useless and causes problems with multiproduct support.
They just defined some addresses and names which should be written directly to
board file. A define that is used only in one place is just confusing things to
removing whole files.

mach/board.h is now empty but is left there if there is better use for that.

Signed-off-by: Kari Suvanto <kari.suvanto@broadcom.com>