clock-r8a7373: Set CLK_ENABLE_ON_INIT for IRQC
As drivers/sh/pm_runtime.c is currently deactivated (due to CONFIG_PDC
being on, and an #ifndef CONFIG_PDC in drivers/sh/Makefile), we lack
normal shmobile platform bus PM runtime support, which means that
drivers' CLKDEV_DEV_ID() clocks are not automatically enabled. So the
entries for "renesas_irqc.0" and "e61c0000.interrupt-controller" are
ineffective.

Work around this by setting CLK_ENABLE_ON_INIT - when/if we figure out
how to get the platform bus PM runtime working properly, this can be
removed.

We seem to cope without this patch, because the RT domain is requesting
MSTP507, but we can't rely on that.

Signed-off-by: Kevin Bracey <kbracey@broadcom.com>