shmobile: add SH_PFC_ pinctrl wrapper macros
To avoid having to repeatedly specify the name of the pin controller,
provide SH_PFC_ wrapper macros to use in place of pinctrl PIN_ macros.

This tidies the code, but more importantly provides a way of handling
setup-xxx.c files that might be shared between chips, and handling the
different name of the PFC when it is initialised from device tree.

Change r8a7373 and its board files to use this facility.

Signed-off-by: Kevin Bracey <kbracey@broadcom.com>