Add non-DT device names to powerdomain table
The power domain request for devices fail in a non-DT enviroment due to the
names being different with DT and without DT. This affects i2c currently, but
when every device moves to DT we need to make corresponding change here as well.

The MMCIF name is also updated since there is a mmc DT support is already done
and queued up in review.

Signed-off-by: Santosh Sivaraj <santosh.sivaraj@broadcom.com>
Signed-off-by: Asha M.V <asha.magadi-venkatesha-murthy@broadcom.com>