dev-bt: Refine device registration
Compact the shared BT device registration functions by using
platform_device_register_data. We can efficiently just create automatic
data with the passed parameters, which is then memdup'ed by the core.

Thus overhead is minimised if the functions are not called because the
current board is using DT - we can conceptually support both DT and
non-DT boards in one image.

Signed-off-by: Kevin Bracey <kbracey@broadcom.com>